Refactor result management to support multiple game types and consolidate result logic into a single Result class
